Pizda

+++ PIRIS-labs/DTOs/ATM/BalanceInquiryDto.cs
@@ -2,7 +2,7 @@
 {
   public class BalanceInquiryDto
   {
-    public float Balance { get; set; }
+    public decimal Balance { get; set; }
     public string ClientFullName { get; set; }
     public string Number { get; set; }
   }
