+++ PIRIS-labs/DTOs/Credit/CreditPercentAccountDto.cs
@@ -6,7 +6,7 @@ namespace PIRIS_labs.DTOs.Credit
   {
     public Account PercentAccount { get; set; }
     public Data.Entities.Credit Credit { get; set; }
-    public float CreditPercent { get; set; }
+    public decimal CreditPercent { get; set; }
     public bool Anuity { get; set; }
     public int Months { get; set; }
   }
