+++ PIRIS-labs/DTOs/Credit/CreditPercentsDto.cs
@@ -7,8 +7,8 @@ namespace PIRIS_labs.DTOs.Credit
     public int RowNumber { get; set; }
     public DateTime Date { get; set; }
 
-    public double MainDebth { get; set; }
-    public double PercentDebth { get; set; }
-    public double PaymentSum { get; set; }
+    public decimal MainDebth { get; set; }
+    public decimal PercentDebth { get; set; }
+    public decimal PaymentSum { get; set; }
   }
 }
