+++ PIRIS-labs/DTOs/Deposit/DepositPercentAccountDto.cs
@@ -4,7 +4,7 @@ namespace PIRIS_labs.DTOs.Deposit
 {
   public class DepositPercentAccountDto
   {
-    public float Percent { get; set; }
+    public decimal Percent { get; set; }
     public Account Account { get; set; }
     public Data.Entities.Deposit Deposit { get; set; }
   }
